Remove usage license restriction anti-feature
It is rather stupid to limit the potential of software intentionally, especially in AGPLv3 software, so I made these light changes to remove the anti-features that are hopefully maintainable across versions of the software. I am not sure what each of the license parameters mean, so you're welcome to comment if you hit any more anti-feature so I can adjust the code. Ideally the project owners can document each of the license parameters so we can make a proper patch to render the licensing system inoperative.